fix(production): restore permission gate on MCP job scheduling
Security hotfix for the MCP production write tools (audit findings 1-3):

1. scheduleJob lost its authorization check. It was rerouted from the
   `schedule` edge function (which enforces `requirePermissions({ update:
   "production" })`) to `triggerJobSchedule`, an Inngest trigger with no gate
   of its own — every ERP route gates on production update before calling it.
   Any authenticated MCP caller could reschedule jobs without the permission.
   Re-apply the gate: scheduleJob now re-checks production update before firing
   the event. It moves into production.mcp.server.ts because the check needs
   getUserClaims (a server-only module) and the service file is client-reachable.

2. production_triggerJobSchedule is itself an ungated MCP tool (the raw
   primitive). Added to MCP_BLOCKED_TOOL_NAMES so the wrapped, gated
   scheduleJob is the only scheduling entry point over MCP.

3. completeJob's permission check missed the "0" all-companies wildcard that
   requirePermissions honors, wrongly denying all-company admins. Both checks
   now share assertProductionPermission, which honors "0".

Verified: `pnpm run build:erp` 4/4; metadata regenerated (production 205 tools,
triggerJobSchedule removed, scheduleJob params intact with mode/direction enums).
