+++ apps/erp/app/routes/api+/mcp+/lib/mcp-blocked-tools.ts
@@ -11,7 +11,12 @@ export const MCP_BLOCKED_TOOL_NAMES: readonly string[] = [
   // Their args require a userId the MCP executor cannot inject (AuthField has
   // no such payload field), so direct calls would only ever fail validation.
   "production_returnPickedRemaindersForOperation",
-  "production_returnPickedRemaindersForJob"
+  "production_returnPickedRemaindersForJob",
+  // Ungated scheduling primitive: it fires the `schedule-job` Inngest event with
+  // no permission check of its own (every ERP route gates on `production` update
+  // before calling it). `production_scheduleJob` is the intended MCP entry point —
+  // it re-applies that gate — so the raw trigger must not be reachable via MCP.
+  "production_triggerJobSchedule"
 ];
 
 export function isMcpBlockedTool(name: string): boolean {
