+++ apps/erp/app/routes/api+/mcp+/lib/tool-metadata.json
@@ -1,6 +1,6 @@
 {
-  "generated": "2026-08-25T15:30:04.284Z",
-  "totalTools": 1448,
+  "generated": "2026-08-25T17:27:47.821Z",
+  "totalTools": 1447,
   "modules": 15,
   "tools": [
     {
@@ -23338,38 +23338,6 @@
       }
     },
     {
-      "name": "production_triggerJobSchedule",
-      "module": "production",
-      "classification": "WRITE",
-      "description": "trigger job schedule",
-      "paramCount": 3,
-      "serviceParams": [
-        "jobId",
-        "companyId",
-        "userId",
-        "mode",
-        "direction"
-      ],
-      "injectAuth": [
-        "companyId"
-      ],
-      "schema": {
-        "type": "object",
-        "properties": {
-          "jobId": {
-            "type": "string"
-          },
-          "mode": {},
-          "direction": {}
-        },
-        "required": [
-          "jobId",
-          "mode",
-          "direction"
-        ]
-      }
-    },
-    {
       "name": "production_getAssemblyInstruction",
       "module": "production",
       "classification": "READ",
@@ -25270,47 +25238,6 @@
       }
     },
     {
-      "name": "production_scheduleJob",
-      "module": "production",
-      "classification": "WRITE",
-      "description": "schedule job",
-      "paramCount": 3,
-      "serviceParams": [
-        "client",
-        "companyId",
-        "userId",
-        "args"
-      ],
-      "injectAuth": [
-        "companyId"
-      ],
-      "schema": {
-        "type": "object",
-        "properties": {
-          "jobId": {
-            "type": "string"
-          },
-          "mode": {
-            "type": "string",
-            "enum": [
-              "initial",
-              "reschedule"
-            ]
-          },
-          "direction": {
-            "type": "string",
-            "enum": [
-              "backward",
-              "forward"
-            ]
-          }
-        },
-        "required": [
-          "jobId"
-        ]
-      }
-    },
-    {
       "name": "production_completeOperation",
       "module": "production",
       "classification": "WRITE",
@@ -25422,6 +25349,47 @@
         "required": [
           "jobId",
           "quantity"
+        ]
+      }
+    },
+    {
+      "name": "production_scheduleJob",
+      "module": "production",
+      "classification": "WRITE",
+      "description": "schedule job",
+      "paramCount": 3,
+      "serviceParams": [
+        "client",
+        "companyId",
+        "userId",
+        "args"
+      ],
+      "injectAuth": [
+        "companyId"
+      ],
+      "schema": {
+        "type": "object",
+        "properties": {
+          "jobId": {
+            "type": "string"
+          },
+          "mode": {
+            "type": "string",
+            "enum": [
+              "initial",
+              "reschedule"
+            ]
+          },
+          "direction": {
+            "type": "string",
+            "enum": [
+              "backward",
+              "forward"
+            ]
+          }
+        },
+        "required": [
+          "jobId"
         ]
       }
     },
